pyproject.toml
[project]
name = "nonebot-plugin-random-stereotypes"
description = "基于Nonebot2的发病语录插件"
dynamic = ["version"]
authors = [
    { name = "Ikaros", email = "327209194@qq.com" },
    { name = "student_2333", email = "lgc2333@126.com" },
]
requires-python = "<4.0,>=3.9"
dependencies = [
    "nonebot2>=2.3.3",
    "nonebot-plugin-alconna>=0.52.1",
    "nonebot-plugin-userinfo>=0.2.6",
    "nonebot-plugin-waiter>=0.7.1",
    "cookit[pyd]>=0.8.1",
]
license = { text = "MIT" }
readme = "README.md"

[project.optional-dependencies]
meme = ["meme-generator>=0.1.4"]
meme-api = ["nonebot-plugin-memes-api>=0.4.11"]

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-random-stereotypes"
repository = "https://github.com/lgc-NB2Dev/nonebot-plugin-random-stereotypes"
documentation = "https://github.com/lgc-NB2Dev/nonebot-plugin-random-stereotypes/blob/master/README.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["nonebot_plugin_random_stereotypes"]

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_random_stereotypes/__init__.py"
